Conflicts:
	plugin/AL_USDMayaTestPlugin/AL/usdmaya/commands/test_ProxyShapeSelect.cpp

+++ lib/AL_USDMaya/AL/usdmaya/nodes/ProxyShapeSelection.cpp
@@ -17,7 +17,6 @@
 #include "AL/usdmaya/nodes/Transform.h"
 #include "AL/usdmaya/nodes/TransformationMatrix.h"
 #include "AL/usdmaya/TypeIDs.h"
-#include "AL/usdmaya/Utils.h"
 #include "AL/usdmaya/Metadata.h"
 #include "AL/usdmaya/DebugCodes.h"
 
@@ -26,12 +25,14 @@
 
 #include <set>
 #include <algorithm>
+#include "AL/usdmaya/utils/Utils.h"
 
 
 namespace AL {
 namespace usdmaya {
 namespace nodes {
 namespace {
+typedef void (*proxy_function_prototype)(void* userData, AL::usdmaya::nodes::ProxyShape* proxyInstance);
 inline void addObjToSelectionList(MSelectionList& list, const MObject& object)
 {
   if(object.hasFn(MFn::kDagNode))
@@ -265,6 +266,8 @@ inline bool ProxyShape::TransformReference::decRef(const TransformReason reason)
     assert(0);
     break;
   }
+
+  TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShapeSelection::TransformReference::decRefEnd %lu %lu %lu\n", m_selected, m_refCount, m_required);
   return !m_required && !m_selected && !m_refCount;
 }
 
@@ -279,6 +282,8 @@ inline void ProxyShape::TransformReference::incRef(const TransformReason reason)
   case kRequired: ++m_required; break;
   default: assert(0); break;
   }
+
+  TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShapeSelection::TransformReference::incRefEnd %lu %lu %lu\n", m_selected, m_refCount, m_required);
 }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -443,7 +448,7 @@ MObject ProxyShape::makeUsdTransformChain(
     uint32_t* createCount,
     MString* resultingPath)
 {
-  TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShapeSelection::makeUsdTransformChainB %s\n", usdPrim.GetPath().GetText());
+  TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShapeSelection::makeUsdTransformChain %s\n", usdPrim.GetPath().GetText());
 
   SdfPath path = usdPrim.GetPath();
   auto iter = m_requiredPaths.find(path);
@@ -543,7 +548,7 @@ MObject ProxyShape::makeUsdTransformChain(
   bool hasMetadata = usdPrim.GetMetadata(Metadata::transformType, &transformType);
   if(hasMetadata && !transformType.empty())
   {
-    node = modifier.createNode(convert(transformType), parentNode);
+    node = modifier.createNode(AL::maya::utils::convert(transformType), parentNode);
     isTransform = false;
     isUsdTransform = false;
     TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShape::makeUsdTransformChain created transformType=%s name=%s\n", transformType.c_str(), usdPrim.GetName().GetText());
@@ -555,7 +560,7 @@ MObject ProxyShape::makeUsdTransformChain(
   }
 
   fn.setObject(node);
-  fn.setName(convert(usdPrim.GetName().GetString()));
+  fn.setName(AL::maya::utils::convert(usdPrim.GetName().GetString()));
 
   //Retrieve the proxy shapes transform path which will be used in the UsdPrim->MayaNode mapping in the case where there is delayed node creation.
   MFnDagNode shapeFn(thisMObject());
@@ -563,9 +568,9 @@ MObject ProxyShape::makeUsdTransformChain(
   MDagPath mayaPath;
   MDagPath::getAPathTo(shapeParent, mayaPath);
   if(resultingPath)
-    *resultingPath = mapUsdPrimToMayaNode(usdPrim, node, &mayaPath);
+    *resultingPath = AL::usdmaya::utils::mapUsdPrimToMayaNode(usdPrim, node, &mayaPath);
   else
-    mapUsdPrimToMayaNode(usdPrim, node, &mayaPath);
+    AL::usdmaya::utils::mapUsdPrimToMayaNode(usdPrim, node, &mayaPath);
 
   if(isUsdTransform)
   {
@@ -578,7 +583,7 @@ MObject ProxyShape::makeUsdTransformChain(
 
     if(modifier2)
     {
-      modifier2->newPlugValueBool(MPlug(node, Transform::pushToPrim()), true);
+      modifier2->newPlugValueBool(ptrNode->pushToPrimPlug(), true);
     }
 
     if(!isTransform)
@@ -598,7 +603,7 @@ MObject ProxyShape::makeUsdTransformChain(
     }
 
     // set the primitive path
-    fileio::translators::DgNodeTranslator::setString(node, Transform::primPath(), path.GetText());
+    modifier.newPlugValueString(ptrNode->primPathPlug(), path.GetText());
   }
   TransformReference ref(node, reason);
   ref.checkIncRef(reason);
@@ -642,7 +647,7 @@ void ProxyShape::makeUsdTransformsInternal(const UsdPrim& usdPrim, const MObject
       UsdPrim prim = *it;
       MObject node = modifier.createNode(Transform::kTypeId, parentNode);
       fn.setObject(node);
-      fn.setName(convert(prim.GetName().GetString()));
+      fn.setName(AL::maya::utils::convert(prim.GetName().GetString()));
       Transform* ptrNode = (Transform*)fn.userNode();
       MPlug inStageData = ptrNode->inStageDataPlug();
       MPlug inTime = ptrNode->timePlug();
@@ -651,11 +656,11 @@ void ProxyShape::makeUsdTransformsInternal(const UsdPrim& usdPrim, const MObject
 
       if(modifier2)
       {
-        modifier2->newPlugValueBool(MPlug(node, Transform::pushToPrim()), true);
+        modifier2->newPlugValueBool(ptrNode->pushToPrimPlug(), true);
       }
 
       // set the primitive path
-      fileio::translators::DgNodeTranslator::setString(node, Transform::primPath(), prim.GetPath().GetText());
+      modifier.newPlugValueString(ptrNode->primPathPlug(), prim.GetPath().GetText());
       TransformReference transformRef(node, reason);
       transformRef.incRef(reason);
       m_requiredPaths.emplace(prim.GetPath(), transformRef);
@@ -682,7 +687,8 @@ void ProxyShape::removeUsdTransformChain_internal(
   MObject object = MObject::kNullObj;
   while(parentPrim)
   {
-    auto it = m_requiredPaths.find(parentPrim.GetPath());
+    SdfPath primPath = parentPrim.GetPath();
+    auto it = m_requiredPaths.find(primPath);
     if(it == m_requiredPaths.end())
     {
       return;
@@ -696,6 +702,7 @@ void ProxyShape::removeUsdTransformChain_internal(
         modifier.reparentNode(object);
         modifier.deleteNode(object);
       }
+      m_currentLockedPrims.erase(primPath);
     }
 
     parentPrim = parentPrim.GetParent();
@@ -710,9 +717,10 @@ void ProxyShape::removeUsdTransformChain(
 {
   TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShapeSelection::removeUsdTransformChain\n");
   SdfPath parentPrim = path;
-  TF_DEBUG(ALUSDMAYA_TRANSLATORS).Msg("ProxyShape::removeUsdTransformChain %s\n", path.GetText());
+  TF_DEBUG(ALUSDMAYA_SELECTION).Msg("ProxyShape::removeUsdTransformChain %s\n", path.GetText());
   MObject parentTM = MObject::kNullObj;
   MObject object = MObject::kNullObj;
+
   while(!parentPrim.IsEmpty())
   {
     auto it = m_requiredPaths.find(parentPrim);
@@ -720,14 +728,23 @@ void ProxyShape::removeUsdTransformChain(
     {
       return;
     }
+
     if(it->second.decRef(reason))
     {
       MObject object = it->second.node();
       if(object != MObject::kNullObj)
       {
-        modifier.reparentNode(object);
-        modifier.deleteNode(object);
+        MObjectHandle h = object;
+
+        // The Xform of the shape may have already been deleted when the shape was deleted
+        if(h.isAlive() && h.isValid())
+        {
+          modifier.reparentNode(object);
+          modifier.deleteNode(object);
+        }
       }
+
+      m_currentLockedPrims.erase(parentPrim);
       m_requiredPaths.erase(it);
     }
 
@@ -875,6 +892,7 @@ void SelectionUndoHelper::doIt()
     MGlobal::setActiveSelectionList(m_newSelection, MGlobal::kReplaceList);
   }
   m_proxy->m_pleaseIgnoreSelection = false;
+  m_proxy->constructLockPrims();
 }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -892,6 +910,7 @@ void SelectionUndoHelper::undoIt()
     MGlobal::setActiveSelectionList(m_previousSelection, MGlobal::kReplaceList);
   }
   m_proxy->m_pleaseIgnoreSelection = false;
+  m_proxy->constructLockPrims();
 }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -1235,11 +1254,14 @@ bool ProxyShape::doSelect(SelectionUndoHelper& helper, const SdfPathVector& orde
 
   if(newlySelectedPaths.length())
   {
+    triggerEvent("PreSelectionChanged");
+
     MPxCommand::setResult(newlySelectedPaths);
+
+    triggerEvent("PostSelectionChanged");
   }
 
   m_pleaseIgnoreSelection = false;
-
   return true;
 }
